+++ unify-front-end/supabase/migrations/20260320_fix_post_metadata_comment_counts.sql
@@ -1,6 +1,6 @@
 create or replace function public.get_post_metadata_batch(post_ids bigint[])
 returns table (
-  post_id integer,
+  post_id bigint,
   like_count integer,
   comment_count integer,
   is_liked boolean,
@@ -12,7 +12,7 @@ security definer
 set search_path = public
 as $$
   with requested_posts as (
-    select distinct unnest(post_ids)::integer as post_id
+    select distinct unnest(post_ids) as post_id
   ),
   post_rows as (
     select
